Move future helper functions to FutureUtil, alias in syntax #40
